Relocate CLI manual utilities into manual module
Move src/cli/lib/manual-utils.js into src/cli/lib/manual/utils.js so the manual tooling lives under a dedicated directory instead of the general lib root. Manual-specific helpers were previously mixed with unrelated CLI utilities, making it harder to see the boundaries around manual features. Update all CLI commands, libraries, and tests to reference the new path.
