+++ src/cli/lib/manual/utils.js
@@ -5,11 +5,11 @@ import {
     assertNonEmptyString,
     parseJsonWithContext,
     toTrimmedString
-} from "./shared-deps.js";
-import { formatDuration } from "./time-utils.js";
-import { formatBytes } from "./byte-format.js";
-import { isNonEmptyArray } from "./shared/array-utils.js";
-import { writeManualFile } from "./manual-file-helpers.js";
+} from "../shared-deps.js";
+import { formatDuration } from "../time-utils.js";
+import { formatBytes } from "../byte-format.js";
+import { isNonEmptyArray } from "../shared/array-utils.js";
+import { writeManualFile } from "../manual-file-helpers.js";
 
 const MANUAL_REPO_ENV_VAR = "GML_MANUAL_REPO";
 const DEFAULT_MANUAL_REPO = "YoYoGames/GameMaker-Manual";
